+++ plugins/main/server/lib/generate-alerts/types.d.ts
@@ -47,7 +47,7 @@ export interface Location {
 }
 
 export interface Field {
-  timestamp: string;
+  '@timestamp': string;
 }
 
 export type GeoLocation = {